Fix critical reader bugs: webtoon scroll, decode safety, memory handling (#20)
- WebtoonRecyclerView: guard isFixingScroll with try/finally so a throwing
  adjustScroll() cannot permanently disable scroll fixing
- WebtoonHolder: make saved scroll nullable so a saved scroll of 0 is not
  confused with "no saved scroll"; clear pending scroll on recycle
- WebtoonImageView: compute scroll range in Double to avoid Int overflow on
  very tall strips
- PageHolder/ReversedPageHolder: do not call resetScaleAndCenter() in
  FIT_CENTER when image dimensions are not yet valid
- ReaderSettings: compare both decoder factories when applying bitmap config
- NativeJpegDecoder: honor requested bitmap config, validate probed
  dimensions, throw ImageDecodeException instead of RuntimeException
- PageViewModel: report conversion failure as the primary error
- BasePageHolder: do not reload immediately after a memory trim while paused
- RegionBitmapDecoder: validate Coil size sentinels, clamp crop region to
  source bounds, fix inverted aspect formula in the wide-image branch
- EdgeDetector: fix inverted edge detection, populate result cache, replace
  per-pixel getPixel() with one bulk getPixels() read
- DoublePageHolder: document intentional zoom-mode behavior

+++ app/src/main/kotlin/org/koharu/miyo/reader/ui/pager/webtoon/WebtoonImageView.kt
@@ -59,7 +59,10 @@ class WebtoonImageView @JvmOverloads constructor(
 		if (!isReady || sWidth == 0 || sHeight == 0 || width == 0) {
 			return 0
 		}
-		val totalHeight = (sHeight * width / sWidth.toFloat()).roundToInt()
+		// Multiply in Double: sHeight * width is evaluated in Int and can
+		// overflow for very tall strips before the division brings the value
+		// back into range.
+		val totalHeight = (sHeight.toDouble() * width / sWidth.toDouble()).roundToInt()
 		// roundToInt() can saturate to Int.MAX_VALUE for absurdly large
 		// sHeight; clamp negatives and the saturation sentinel.
 		if (totalHeight <= 0 || totalHeight == Int.MAX_VALUE) {
